Rename Manipulation folder to Processor
