+++ qt/scientific_interfaces/Inelastic/Processor/DataManipulation.cpp
