+++ qt/scientific_interfaces/Inelastic/Processor/DataManipulationInterface.ui
