+++ qt/scientific_interfaces/Inelastic/Processor/ElwinPresenter.cpp
