+++ qt/scientific_interfaces/Inelastic/Processor/ElwinView.cpp
