+++ qt/scientific_interfaces/Inelastic/Processor/IElwinView.h
