+++ qt/scientific_interfaces/Inelastic/Processor/IIqtView.h
