+++ qt/scientific_interfaces/Inelastic/Processor/ISqwView.h
