+++ qt/scientific_interfaces/Inelastic/Processor/ISymmetriseView.h
