+++ qt/scientific_interfaces/Inelastic/Processor/IqtModel.cpp
