+++ qt/scientific_interfaces/Inelastic/Processor/IqtView.h
