+++ qt/scientific_interfaces/Inelastic/Processor/MomentsModel.cpp
