+++ qt/scientific_interfaces/Inelastic/Processor/MomentsModel.h
