+++ qt/scientific_interfaces/Inelastic/Processor/MomentsTab.ui
