+++ qt/scientific_interfaces/Inelastic/Processor/MomentsView.cpp
