+++ qt/scientific_interfaces/Inelastic/Processor/SqwModel.cpp
