+++ qt/scientific_interfaces/Inelastic/Processor/SqwPresenter.h
