+++ qt/scientific_interfaces/Inelastic/Processor/SqwTab.ui
