+++ qt/scientific_interfaces/Inelastic/Processor/SqwView.cpp
