+++ qt/scientific_interfaces/Inelastic/Processor/SymmetriseModel.h
