+++ qt/scientific_interfaces/Inelastic/Processor/SymmetrisePresenter.h
