+++ qt/scientific_interfaces/Inelastic/Processor/SymmetriseTab.ui
