+++ qt/scientific_interfaces/Inelastic/Processor/SymmetriseView.h
