+++ qt/scientific_interfaces/Inelastic/test/Processor/SymmetriseModelTest.h
@@ -6,11 +6,11 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "Manipulation/SymmetriseModel.h"
 #include "MantidAPI/AnalysisDataService.h"
 #include "MantidAPI/ITableWorkspace.h"
 #include "MantidAPI/TableRow.h"
 #include "MantidFrameworkTestHelpers/WorkspaceCreationHelper.h"
+#include "Processor/SymmetriseModel.h"
 #include <cxxtest/TestSuite.h>
 #include <gmock/gmock.h>
